Cleaning up code, adding CrlUtils static class so the code is reusable
