--- X509-Testsuite/src/main/java/de/rub/nds/x509anvil/suite/tests/util/CrlUtils.java
@@ -1,5 +1,0 @@
-package de.rub.nds.x509anvil.suite.tests.util;
-
-public class CrlUtils {
-
-}
